temp/Python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
python-multipart==0.0.9
google-generativeai==0.8.3
statsmodels==0.14.2  # This line is critical for ARIMA